# Optimize column vector filter