--- a/dbms/src/Columns/ColumnVector.cpp
+++ b/dbms/src/Columns/ColumnVector.cpp
@@ -206,67 +206,6 @@ void ColumnVector<T>::insertRangeFrom(const IColumn & src, size_t start, size_t 
 namespace
 {
 
-/// Transform 64-byte mask to 64-bit mask
-inline UInt64 bytes64MaskToBits64Mask(const UInt8 * bytes64)
-{
-#if defined(__AVX512F__) && defined(__AVX512BW__)
-    const __m512i vbytes = _mm512_loadu_si512(reinterpret_cast<const void *>(bytes64));
-    UInt64 res = _mm512_testn_epi8_mask(vbytes, vbytes);
-#elif defined(__AVX__) && defined(__AVX2__)
-    const __m256i zero32 = _mm256_setzero_si256();
-    UInt64 res = (static_cast<UInt64>(_mm256_movemask_epi8(_mm256_cmpeq_epi8(
-                      _mm256_loadu_si256(reinterpret_cast<const __m256i *>(bytes64)),
-                      zero32)))
-                  & 0xffffffff)
-        | (static_cast<UInt64>(_mm256_movemask_epi8(_mm256_cmpeq_epi8(
-               _mm256_loadu_si256(reinterpret_cast<const __m256i *>(bytes64 + 32)),
-               zero32)))
-           << 32);
-#elif defined(__SSE2__)
-    const __m128i zero16 = _mm_setzero_si128();
-    UInt64 res = (static_cast<UInt64>(_mm_movemask_epi8(_mm_cmpeq_epi8(
-                      _mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64)),
-                      zero16)))
-                  & 0xffff)
-        | ((static_cast<UInt64>(_mm_movemask_epi8(_mm_cmpeq_epi8(
-                _mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64 + 16)),
-                zero16)))
-            << 16)
-           & 0xffff0000)
-        | ((static_cast<UInt64>(_mm_movemask_epi8(_mm_cmpeq_epi8(
-                _mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64 + 32)),
-                zero16)))
-            << 32)
-           & 0xffff00000000)
-        | ((static_cast<UInt64>(_mm_movemask_epi8(_mm_cmpeq_epi8(
-                _mm_loadu_si128(reinterpret_cast<const __m128i *>(bytes64 + 48)),
-                zero16)))
-            << 48)
-           & 0xffff000000000000);
-#elif defined(__aarch64__) && defined(__ARM_NEON)
-    const uint8x16_t bitmask = {0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80, 0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80};
-    const auto * src = reinterpret_cast<const unsigned char *>(bytes64);
-    const uint8x16_t p0 = vceqzq_u8(vld1q_u8(src));
-    const uint8x16_t p1 = vceqzq_u8(vld1q_u8(src + 16));
-    const uint8x16_t p2 = vceqzq_u8(vld1q_u8(src + 32));
-    const uint8x16_t p3 = vceqzq_u8(vld1q_u8(src + 48));
-    uint8x16_t t0 = vandq_u8(p0, bitmask);
-    uint8x16_t t1 = vandq_u8(p1, bitmask);
-    uint8x16_t t2 = vandq_u8(p2, bitmask);
-    uint8x16_t t3 = vandq_u8(p3, bitmask);
-    uint8x16_t sum0 = vpaddq_u8(t0, t1);
-    uint8x16_t sum1 = vpaddq_u8(t2, t3);
-    sum0 = vpaddq_u8(sum0, sum1);
-    sum0 = vpaddq_u8(sum0, sum0);
-    UInt64 res = vgetq_lane_u64(vreinterpretq_u64_u8(sum0), 0);
-#else
-    UInt64 res = 0;
-    for (size_t i = 0; i < 64; ++i)
-        res |= static_cast<UInt64>(0 == bytes64[i]) << i;
-#endif
-    return ~res;
-}
-
 /// If mask is a number of this kind: [0]*[1]* function returns the length of the cluster of 1s.
 /// Otherwise it returns the special value: 0xFF.
 UInt8 prefixToCopy(UInt64 mask)
@@ -297,7 +236,7 @@ inline void doFilterAligned(const UInt8 *& filt_pos, const UInt8 *& filt_end_ali
 {
     while (filt_pos < filt_end_aligned)
     {
-        UInt64 mask = bytes64MaskToBits64Mask(filt_pos);
+        UInt64 mask = ToBits64(filt_pos);
         const UInt8 prefix_to_copy = prefixToCopy(mask);
 
         if (0xFF != prefix_to_copy)