### What problem does this PR solve?

Issue Number: ref #6839

Problem Summary:

### What is changed and how it works?

rewrite `ColumnVector<T>::filter` with `avx2`

perf test as follows (larger is better):

```shell
[qiuyang@172-16-5-85 ~/projects/tiflash/cmake-build-release 11:55:31]$ ./dbms/src/Columns/tests/column_vector_perftest int64 filter 10000 1 30
Test int64-filter rows=10000 columns=1 seconds=30
FilterV1: 117192    
FilterV2: 157265    
FilterV3: 157951    
[qiuyang@172-16-5-85 ~/projects/tiflash/cmake-build-release 11:56:09]$ ./dbms/src/Columns/tests/column_vector_perftest int64 filter 10000 10 30
Test int64-filter rows=10000 columns=10 seconds=30
FilterV1: 157057    
FilterV2: 290671    
FilterV3: 95683     
[qiuyang@172-16-5-85 ~/projects/tiflash/cmake-build-release 11:56:54]$ ./dbms/src/Columns/tests/column_vector_perftest int64 filter 10000 50 30
Test int64-filter rows=10000 columns=50 seconds=30
FilterV1: 408640    
FilterV2: 803905    
FilterV3: 398920    
[qiuyang@172-16-5-85 ~/projects/tiflash/cmake-build-release 11:57:32]$ ./dbms/src/Columns/tests/column_vector_perftest int64 filter 10000 100 30
Test int64-filter rows=10000 columns=100 seconds=30
FilterV1: 532940    
FilterV2: 983276    
FilterV3: 582252    
[qiuyang@172-16-5-85 ~/projects/tiflash/cmake-build-release 11:58:07]$ ./dbms/src/Columns/tests/column_vector_perftest int64 filter 10000 101 30
Test int64-filter rows=10000 columns=101 seconds=30
FilterV1: 833012    
FilterV2: 1339852   
FilterV3: 1159898
```

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```